pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prime_plot"
version = "0.1.0"
description = "GPU/NPU-accelerated prime number visualization and pattern recognition"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Prime Plot Contributors"}
]
keywords = ["prime numbers", "ulam spiral", "visualization", "machine learning", "gpu"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "pillow>=9.0.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
fast = [
    "primesieve>=2.3.0",
]
gpu = [
    "cupy-cuda12x>=12.0.0",
]
npu = [
    "openvino>=2023.0.0",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
all = [
    "prime_plot[fast,gpu,npu,dev]",
]

[project.scripts]
prime-plot = "prime_plot.cli:main"

[project.urls]
Repository = "https://github.com/yourusername/prime_plot"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "W", "I", "UP", "B", "C4"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
